ai-services/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
numpy==1.26.4
pandas==2.2.2
opencv-python-headless==4.9.0.80
tensorflow==2.16.1
scikit-learn==1.4.2
python-dotenv==1.0.1
pydantic==2.7.1
